+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * `defianno` tool added (helps with annotating definienda)
 * `snify`'s `x𝑖` command now also shows the import path of symbols (if in scope)
 * Can quit in session dialog
+* You can override `snify`'s language detection with the `--lang` argument
 
 **Other modifications/improvements:**